# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/operator/work-queue": {
      "get": {
        "operationId": "listWorkQueue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperatorWorkQueueItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "List review work without sensitive evidence data",
        "tags": ["operator"]
      }
    },
...
    "/v1/operator/sessions/{sessionId}/review": {
      "get": {
        "operationId": "getSessionEvidenceReview",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionEvidenceReviewResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Get a privacy-minimized, server-bound session evidence review",
        "tags": ["operator"]
      }
    },
    "/v1/operator/sessions/{sessionId}/reject": {
      "post": {
        "operationId": "rejectSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RejectSessionDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Reject a reviewed workout session",
        "tags": ["operator"]
      }
    },
    "/v1/operator/sessions/{sessionId}/verify": {
      "post": {
        "operationId": "verifySession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifySessionDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Manually verify a reviewed workout session",
        "tags": ["operator"]
      }
    },
    "/v1/operator/draws/lock": {
      "post": {
        "operationId": "lockDraw",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LockDrawDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Lock a competition draw snapshot and commitment",
        "tags": ["operator"]
      }
    },
    "/v1/operator/draws/{drawId}/settle": {
      "post": {
        "operationId": "settleDraw",
        "parameters": [
          {
            "name": "drawId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SettleDrawDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Reveal the draw seed and settle winners",
        "tags": ["operator"]
      }
    },
    "/v1/operator/region-verifications/{verificationId}/decision": {
      "post": {
        "operationId": "decideRegion",
        "parameters": [
          {
            "name": "verificationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideRegionVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Approve or reject a region-verification review",
        "tags": ["operator"]
      }
    },
    "/v1/operator/partner-applications/{applicationId}/decision": {
      "post": {
        "operationId": "decidePartner",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecidePartnerApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Review a partner application",
        "tags": ["operator"]
      }
    },
    "/v1/operator/privacy-requests/{privacyRequestId}/decision": {
      "post": {
        "operationId": "decidePrivacy",
        "parameters": [
          {
            "name": "privacyRequestId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecidePrivacyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Start or reject a privacy request",
        "tags": ["operator"]
      }
    },
    "/v1/operator/profile-media/{mediaId}/review-action": {
      "get": {
        "operationId": "getProfileMediaReviewAction",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileMediaReviewActionDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Create a short-lived private avatar review action",
        "tags": ["operator"]
      }
    },
    "/v1/operator/profile-media/{mediaId}/decision": {
      "post": {
        "operationId": "decideProfileMedia",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DecideProfileMediaDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorActionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Approve or reject moderated avatar media",
        "tags": ["operator"]
      }
    },
...
    "/v1/operator/payout-claims/{claimId}/approve": {
      "post": {
        "operationId": "approve",
        "parameters": [
          {
            "name": "claimId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorPayoutActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorPayoutClaimResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Approve a reviewed winner claim for payee setup",
        "tags": ["operator payouts"]
      }
    },
...
  },
  "info": {
    "title": "GoGymGo API",
    "description": "Authoritative API for GoGymGo accounts, verified competition activity, ledgers, results, and payouts.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "firebase": {
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "type": "http",
        "description": "Firebase ID token verified by the GoGymGo API."
      },
      "hyperwallet-webhook": {
        "type": "http",
        "scheme": "basic",
        "description": "Dedicated credentials configured only for Hyperwallet."
      }
    },
    "schemas": {
...
      "OperatorWorkQueueItemDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": {
            "type": "string",
            "enum": [
              "partner_application",
              "payout_claim",
              "privacy_request",
              "profile_media",
              "region_verification",
              "workout_session"
            ]
          },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "kind", "status", "createdAt"]
      },
...
      "DeviceAttestationReviewDto": {
        "type": "object",
        "properties": {
          "count": { "type": "number", "minimum": 0 },
          "minimumRequiredCount": { "type": "number", "minimum": 0 },
          "required": { "type": "boolean" },
          "trustStates": { "type": "array", "items": { "type": "string" } },
          "uniqueTokenCount": { "type": "number", "minimum": 0 }
        },
        "required": [
          "count",
          "minimumRequiredCount",
          "required",
          "trustStates",
          "uniqueTokenCount"
        ]
      },
      "FaceCheckReviewDto": {
        "type": "object",
        "properties": {
          "count": { "type": "number", "minimum": 0 },
          "minimumRequiredCount": { "type": "number", "minimum": 0 },
          "required": { "type": "boolean" },
          "trustStates": { "type": "array", "items": { "type": "string" } },
          "maximumConfidence": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "nullable": true
          }
        },
        "required": [
          "count",
          "minimumRequiredCount",
          "required",
          "trustStates",
          "maximumConfidence"
        ]
      },
      "GymQrReviewDto": {
        "type": "object",
        "properties": {
          "count": { "type": "number", "minimum": 0 },
          "minimumRequiredCount": { "type": "number", "minimum": 0 },
          "required": { "type": "boolean" },
          "trustStates": { "type": "array", "items": { "type": "string" } },
          "uniquePayloadCount": { "type": "number", "minimum": 0 }
        },
        "required": [
          "count",
          "minimumRequiredCount",
          "required",
          "trustStates",
          "uniquePayloadCount"
        ]
      },
      "HeartRateReviewDto": {
        "type": "object",
        "properties": {
          "count": { "type": "number", "minimum": 0 },
          "minimumRequiredCount": { "type": "number", "minimum": 0 },
          "required": { "type": "boolean" },
          "trustStates": { "type": "array", "items": { "type": "string" } },
          "averageBpm": { "type": "number", "nullable": true },
          "maximumBpm": { "type": "number", "nullable": true },
          "minimumBpm": { "type": "number", "nullable": true }
        },
        "required": [
          "count",
          "minimumRequiredCount",
          "required",
          "trustStates",
          "averageBpm",
          "maximumBpm",
          "minimumBpm"
        ]
      },
      "SessionEvidenceReviewGroupsDto": {
        "type": "object",
        "properties": {
          "deviceAttestation": {
            "$ref": "#/components/schemas/DeviceAttestationReviewDto"
          },
          "faceCheck": { "$ref": "#/components/schemas/FaceCheckReviewDto" },
          "gymQr": { "$ref": "#/components/schemas/GymQrReviewDto" },
          "heartRate": { "$ref": "#/components/schemas/HeartRateReviewDto" }
        },
        "required": ["deviceAttestation", "faceCheck", "gymQr", "heartRate"]
      },
      "SessionEvidenceReviewResponseDto": {
        "type": "object",
        "properties": {
          "competitionId": { "type": "string", "format": "uuid" },
          "completedAt": { "type": "string", "format": "date-time" },
          "durationMinutes": { "type": "number", "minimum": 0 },
          "eligibleDate": { "type": "string", "format": "date" },
          "evidence": {
            "$ref": "#/components/schemas/SessionEvidenceReviewGroupsDto"
          },
          "evidenceSnapshotSha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64
          },
          "limitations": { "type": "array", "items": { "type": "string" } },
          "minimumDurationMinutes": { "type": "number", "minimum": 0 },
          "policyVersion": { "type": "string" },
          "sessionId": { "type": "string", "format": "uuid" },
          "startedAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string" }
        },
        "required": [
          "competitionId",
          "completedAt",
          "durationMinutes",
          "eligibleDate",
          "evidence",
          "evidenceSnapshotSha256",
          "limitations",
          "minimumDurationMinutes",
          "policyVersion",
          "sessionId",
          "startedAt",
          "status"
        ]
      },
      "SessionEvidenceFindingsDto": {
        "type": "object",
        "properties": {
          "deviceAttestation": {
            "type": "string",
            "enum": ["approved", "not_required", "rejected"]
          },
          "faceCheck": {
            "type": "string",
            "enum": ["approved", "not_required", "rejected"]
          },
          "gymQr": {
            "type": "string",
            "enum": ["approved", "not_required", "rejected"]
          },
          "heartRate": {
            "type": "string",
            "enum": ["approved", "not_required", "rejected"]
          }
        },
        "required": ["deviceAttestation", "faceCheck", "gymQr", "heartRate"]
      },
      "RejectSessionDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "evidenceSnapshotSha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64
          },
          "findings": {
            "$ref": "#/components/schemas/SessionEvidenceFindingsDto"
          }
        },
        "required": ["reason", "evidenceSnapshotSha256", "findings"]
      },
      "OperatorActionResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": { "type": "string" }
        },
        "required": ["id", "status"]
      },
      "VerifySessionDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "evidenceSnapshotSha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64
          },
          "findings": {
            "$ref": "#/components/schemas/SessionEvidenceFindingsDto"
          }
        },
        "required": ["reason", "evidenceSnapshotSha256", "findings"]
      },
      "LockDrawDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "competitionId": { "type": "string", "format": "uuid" },
          "seedCommitment": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64
          }
        },
        "required": ["reason", "competitionId", "seedCommitment"]
      },
      "SettleDrawDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "seedReveal": { "type": "string", "maxLength": 64, "minLength": 64 }
        },
        "required": ["reason", "seedReveal"]
      },
      "DecideRegionVerificationDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "decision": { "type": "string", "enum": ["approved", "rejected"] },
          "expiresAt": { "type": "string", "format": "date-time" }
        },
        "required": ["reason", "decision"]
      },
      "DecidePartnerApplicationDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "decision": {
            "type": "string",
            "enum": ["approved", "in_review", "rejected"]
          }
        },
        "required": ["reason", "decision"]
      },
      "DecidePrivacyRequestDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "decision": { "type": "string", "enum": ["processing", "rejected"] }
        },
        "required": ["reason", "decision"]
      },
      "ProfileMediaReviewActionDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "contentType": { "type": "string" },
          "contentLength": { "type": "number" },
          "submittedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "url": { "type": "string", "format": "uri" }
        },
        "required": [
          "id",
          "contentType",
          "contentLength",
          "submittedAt",
          "expiresAt",
          "url"
        ]
      },
      "DecideProfileMediaDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "decision": { "type": "string", "enum": ["approved", "rejected"] }
        },
        "required": ["reason", "decision"]
      },
...
      "OperatorPayoutActionDto": {
        "type": "object",
        "properties": {
          "expectedVersion": { "type": "number", "minimum": 1 },
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 }
        },
        "required": ["expectedVersion", "reason"]
      },
      "OperatorPayoutClaimResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": [
              "pending_review",
              "action_required",
              "verification_pending",
              "ready",
              "processing",
              "paid",
              "failed",
              "cancelled"
            ]
          },
          "version": { "type": "number" }
        },
        "required": ["id", "status", "version"]
      },
...
    }
  }
}